requirements.txt
flask==2.3.3
flask-cors==4.0.0
stripe==5.5.0
python-dotenv==1.0.0
flask-limiter==3.5.0
flask-talisman==1.1.0
cryptography==41.0.7
pyjwt==2.8.0